Credential and Microledger terminology updates
I replaced 'credential' with the more accurate term 'proof', and replaced microledgers with the term Peer DID. I explained this probably in too much detail in the section near the bottom of the paper, but wanted to be clear about what I meant by Peer DID.

The term Peer DID is a suggestion here and has not been ratified. Replacement term suggestions are welcome.

Signed-off-by: Brent <brent.zundel@gmail.com>